revamping the message service
